+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added power scaler for new board.
 - Added Main Index link to develop.html.
 - Added feature to automatically reconnect BLE devices if both are specified.
+- Added resistance capture to Echelon.
+- Added Resistance capture to Flywheel.
+- Added Resistance Capture to Peloton.
+- Added Resistance capture to FTMS.
 
 ### Changed
 - PowerTable values are now adjusted to 90 RPM cad on input.